Reorganizing tests
Each object now has its own file.

Also added some Epoch.merge() tests.
